+++ src/core/compilers/vue2.ts
