+++ src/core/compilers/vue3.ts
